+++ src/user/user_model.cc
@@ -929,7 +929,7 @@ void mjCModel::SetSizes(void) {
       nu++;
     } else {
       nu++;
-      na++;
+      na += actuators[i]->actdim;
     }
   }
 
@@ -1834,7 +1834,9 @@ void mjCModel::CopyObjects(mjModel* m) {
     m->actuator_biastype[i] = pac->biastype;
     m->actuator_trnid[2*i] = pac->trnid[0];
     m->actuator_trnid[2*i+1] = pac->trnid[1];
-    m->actuator_actadr[i] = pac->dyntype == mjDYN_NONE ? -1 : adr++;
+    m->actuator_actadr[i] = pac->dyntype == mjDYN_NONE ? -1 : adr;
+    adr += pac->actdim;
+    m->actuator_actnum[i] = pac->actdim;
     m->actuator_group[i] = pac->group;
     m->actuator_ctrllimited[i] = pac->ctrllimited;
     m->actuator_forcelimited[i] = pac->forcelimited;