+++ unity/Runtime/Bindings/MujocoBindings.cs
@@ -2132,6 +2132,7 @@ public unsafe struct mjModel_ {
   public int* actuator_biastype;
   public int* actuator_trnid;
   public int* actuator_actadr;
+  public int* actuator_actnum;
   public int* actuator_group;
   public byte* actuator_ctrllimited;
   public byte* actuator_forcelimited;